add song and additional guests fields to rsvp form
