+++ components/Form/GuestForm/GuestForm.js
@@ -11,6 +11,7 @@ import {
 } from '../';
 import { codeGenerator } from '../../../utils/utils';
 import AdditionalGuests from '../AdditionalGuests';
+import Song from '../Song';
 
 const FormField = ({ children }) => <Box>{children}</Box>;
 
@@ -100,23 +101,7 @@ const GuestForm = ({ data, apiCall, submitText }) => {
                 />
               </FormField>
 
-              <FormField>
-                <FormLabel sx={{ marginBottom: 2 }} component='legend'>
-                  Song Suggestion
-                </FormLabel>
-
-                <Grid container spacing={1}>
-                  <Grid item xs={12} sm={6}>
-                    <FormikTextField name='songRec.title' label='Title' />
-                  </Grid>
-                  <Grid item xs={12} sm={6}>
-                    <FormikTextField
-                      name='songRec.artist'
-                      label='Group/Artist'
-                    />
-                  </Grid>
-                </Grid>
-              </FormField>
+              <Song />
 
               <Divider />
