+++ components/Form/Song/index.js
@@ -1,0 +1,3 @@
+import Song from './Song';
+
+export default Song;
